requirements.txt
pyyaml>=5.1                 # mit
pycdlib                     # lgpl
oslo.concurrency            # apache2
jinja2<3.0,>=2.10.1         # bsd
setproctitle                # bsd
click<8.0,>=7.1.1           # bsd
prettytable                 # bsd
tox                         # mit
flake8                      # mit
testtools                   # mit
Werkzeug<2.0                # bsd
flask<2.0                   # bsd
flask_restful               # bsd
psutil                      # bsd
prometheus_client           # apache2
etcd3                       # apache2
etcd3gw==0.2.6              # apache2
flask-jwt-extended==3.25.0  # mit
bcrypt                      # apache2
gunicorn                    # mit
pylogrus                    # mit
pydantic                    # mit
python-magic                # mit
flask-request-id-middleware # mit
webargs                     # mit

# Is difficult to get install working, use system packages instead. On Ubuntu
# those are: libvirt-daemon-system libvirt-dev python3-libvirt
#libvirt-python      # lgpl
